% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Adding Two Numbers Together}
\usage{
add(a, b)
}
\arguments{
\item{a}{the first number}

\item{b}{the second number}
}
\value{
the summation result
}
\description{
This function adds two numbers together in the usual way
}
\examples{

add(2,4)

}
