% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSurvey.R
\name{aggregateSurvey}
\alias{aggregateSurvey}
\title{Aggregate estimators from different surveys.}
\usage{
aggregateSurvey(data)
}
\arguments{
\item{data}{Output from \code{\link{countrySummary_mult}}}
}
\value{
Estimators aggregated across surveys.
}
\description{
Aggregate estimators from different surveys.
}
\examples{
\dontrun{
TODO
}

}
