% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothArea.R
\name{compareEstimates}
\alias{compareEstimates}
\title{Plot heatmap comparing pairwise posterior exceedence probabilities for all
areas}
\usage{
compareEstimates(x, posterior.sample = NULL)
}
\arguments{
\item{x}{svysae object. Plots are created for all models in this object.}

\item{posterior.sample}{Matrix of posteriors samples of area level quantities with one row for each area and one column for each sample. This argument may be specified to only provide a heatmap for the desired samples.}
}
\value{
ggplot containing heat map of pairwise comparisons
}
\description{
Plot heatmap comparing pairwise posterior exceedence probabilities for all
areas
}
\examples{
\dontrun{
data(DemoData2)
data(DemoMap2)
des0 <- svydesign(ids = ~clustid+id, strata = ~strata,
                  weights = ~weights, data = DemoData2, nest = T)
Xmat <- aggregate(age~region, data = DemoData2, FUN = mean)

cts.res <- smoothArea(tobacco.use ~ 1,
                      domain = ~region,
                      design = des0,
                      adj.mat = DemoMap2$Amat, 
                      pc.u = 1,
                      pc.alpha = 0.01,
                      pc.u.phi = 0.5,
                      pc.alpha.phi = 2/3,
                      return.samples = T)
compareEstimates(cts.res)
}
}
