% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData.R
\docType{data}
\name{DemoData}
\alias{DemoData}
\title{Simulated child mortality person-month dataset.}
\format{A list of with five components, named by survey year.}
\source{
\url{https://dhsprogram.com/data/model-datasets.cfm}
}
\usage{
data(DemoData)
}
\description{
A small simulated dataset with 4 regions and 5 survey years. This  does not represent 
any real country's data and are based on a subset of the model dataset
provided by DHS.
}
\keyword{datasets}
