% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData2.R
\docType{data}
\name{DemoData2}
\alias{DemoData2}
\title{Simulated dataset for prevalence mapping.}
\format{
A data.frame of 7 variables.
}
\source{
\url{https://dhsprogram.com/data/model-datasets.cfm}
}
\usage{
data(DemoData2)
}
\description{
A small fake dataset with 8 regions and two response variables: age and tobacco.use. This  does not represent 
any real country's data and are based on a subset of the model dataset
provided by DHS.
}
\keyword{datasets}
