% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoMap.R
\docType{data}
\name{DemoMap}
\alias{DemoMap}
\title{Map dataset for vignette.}
\format{An object of class \code{list} of length 2.}
\source{
\url{https://spatialdata.dhsprogram.com/boundaries/#view=table&countryId=UG}
}
\usage{
DemoMap
}
\description{
Shapefiles are from 1995 Uganda Admin 1 regions provided by DHS, but the data do not represent real information about any country.
}
\details{
\itemize{
 \item geo. Geographic map files
 \item Amat. Adjacency matrix for regions
}
}
\keyword{datasets}
