% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_RW.R
\name{fit_RW}
\alias{fit_RW}
\title{Function to get RW2 output}
\usage{
fit_RW(data, inla_mod, years, geo, newyear = "15-19", quantiles = c(0.025,
  0.5, 0.975))
}
\arguments{
\item{data}{country summary data from \code{\link{countrySummary_mult}}}

\item{inla_mod}{output from \code{\link{fitINLA}}}

\item{years}{years string vector}

\item{geo}{geographic polygon object}

\item{newyear}{string of years for projection, defaults to \code{'15-19'}}

\item{quantiles}{quantiles desired, defaults to \code{c(0.025,0.5,0.975)}}
}
\value{
Results from RW2 model fit, including projection.
}
\description{
See package vignette for usage details.
}
