% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothSurvey.R
\name{fitGeneric}
\alias{fitGeneric}
\title{Fit space-time smoothing models for a binary outcome from complex surveys.}
\usage{
fitGeneric(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{fitGeneric()} was renamed to \code{smoothSurvey()} to create a more
consistent API.
}
\keyword{internal}
