% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothDirect.R
\name{fitINLA}
\alias{fitINLA}
\title{Smoothed direct estimates for mortality rates}
\usage{
fitINLA(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{fitINLA()} was renamed to \code{smoothDirect()} to create a more
consistent API.
}
\keyword{internal}
