% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothCluster.R
\name{fitINLA2}
\alias{fitINLA2}
\title{Cluster-level space-time smoothing models for mortality rates}
\usage{
fitINLA2(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{fitINLA2()} was renamed to \code{smoothCluster()} to create a more
consistent API.
}
\keyword{internal}
