% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdjusted.R
\name{getAdjusted}
\alias{getAdjusted}
\title{Adjust direct estimates and their associated variances}
\usage{
getAdjusted(data, ratio, time = "years", region = "region",
  est = "mean", logit = "logit.est", logit.var = "var.est",
  logit.prec = "logit.prec", logit.lower = "lower",
  logit.upper = "upper")
}
\arguments{
\item{data}{data frame of the adjusted estimates and the associated uncertainties, see the arguments below for specific columns.}

\item{ratio}{the ratio of unadjusted mortality rates to the true mortality rates. It can be either a data frame with three columns (region, time, and ratio) if adjustment factor differ by region; or a data frame with two columns (time and ratio) if adjustment factor only varies over time.}

\item{time}{the column name for time in the data}

\item{region}{the column name for region in the data}

\item{est}{the column name for unadjusted mortality rates in the data}

\item{logit}{the column name for the logit of the unadjusted mortality rates in the data}

\item{logit.var}{the column name for the variance of the logit of the unadjusted mortality rates in the data}

\item{logit.prec}{the column name for the precision of the logit of the unadjusted mortality rates in the data}

\item{logit.lower}{the column name for the 95\% lower bound of the logit of the unadjusted mortality rates in the data}

\item{logit.upper}{the column name for the 95\% lower bound of the logit of the unadjusted mortality rates in the data}
}
\value{
adjusted dataset of the same columns.
}
\description{
Adjust direct estimates and their associated variances
}
\examples{

 
}
