% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmat.R
\name{getAmat}
\alias{getAmat}
\title{Extract adjacency matrix from the map}
\usage{
getAmat(geo, names)
}
\arguments{
\item{geo}{SpatialPolygonsDataFrame of the map}

\item{names}{character vector of region ids to be added to the neighbours list}
}
\value{
Spatial djacency matrix.
}
\description{
Extract adjacency matrix from the map
}
\examples{
data(DemoMap) 
mat <- getAmat(geo = DemoMap$geo, names = DemoMap$geo$REGNAME)
mat
DemoMap$Amat
 
}
