% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBirths.R
\name{getBirths}
\alias{getBirths}
\title{Function to get Births file from DHS Stata}
\usage{
getBirths(filepath, surveyyear)
}
\arguments{
\item{filepath}{filepath of raw .dta file from DHA}

\item{surveyyear}{year of survey}
}
\value{
A list of birthmonth data
}
\description{
Function to get Births file from DHS Stata
}
\examples{
\dontrun{
my_fp <- "/myExampleFilepath/surveyData.DTA"
Uganda <- getBirths(filepath = my_fp, surveyyear = 2015) 
}

}
