% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiag.R
\name{getDiag}
\alias{getDiag}
\title{Make diagnostic plots}
\usage{
getDiag(inla_mod, field = c("space", "time", "spacetime")[1],
  year_range = c(1985, 2019), year_label = c("85-89", "90-94", "95-99",
  "00-04", "05-09", "10-14", "15-19"), Amat = NULL, CI = 0.95)
}
\arguments{
\item{inla_mod}{output from \code{\link{fitINLA}}}

\item{field}{which random effects to plot. It can be one of the following: space, time, and spacetime.}

\item{year_range}{range corresponding to year label}

\item{year_label}{vector of year string vector}

\item{Amat}{adjacency matrix}

\item{CI}{Desired level of credible intervals}
}
\value{
List of diagnostic plots
}
\description{
Make diagnostic plots
}
\examples{
\dontrun{
# TODO 
} 

}
