% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmoothedDensity.R
\name{getSmoothedDensity}
\alias{getSmoothedDensity}
\title{Function to calculate and plot densities of the projected estimates from INLA for each time and region.}
\usage{
getSmoothedDensity(inla_mod = NULL, results = NULL,
  year_range = c(1985, 2019), year_label = c("85-89", "90-94", "95-99",
  "00-04", "05-09", "10-14", "15-19"), Amat = NULL, nsim = 1000,
  weight.strata = NULL, verbose = FALSE, mc = 0,
  include_time_unstruct = FALSE, year_plot = NULL, byyear = TRUE,
  ncol = 4, per1000 = FALSE, order = 0, ...)
}
\arguments{
\item{inla_mod}{output from \code{\link{smoothDirect}}}

\item{results}{output from \code{\link{getSmoothedDensity}}. This argument can be specified to avoid calculating densities again when only the visualization changes.}

\item{year_range}{range corresponding to year label}

\item{year_label}{vector of year string vector}

\item{Amat}{adjacency matrix}

\item{nsim}{number of simulations}

\item{weight.strata}{a data frame with three columns, years, region, and proportion of each strata for the corresponding time period and region.}

\item{verbose}{logical indicator whether to print progress messages from inla.posterior.sample.}

\item{mc}{number of monte carlo draws to approximate the marginal prevalence/hazards for binomial model. If mc = 0, analytical approximation is used. The analytical approximation is invalid for hazard modeling with more than one age groups.}

\item{include_time_unstruct}{logical indicator whether to include the temporal unstructured effects (i.e., shocks) in the smoothed estimates.}

\item{year_plot}{similar to year_label, a vector indicate which years to plot}

\item{byyear}{logical indicator for whether the output uses years as facets.}

\item{ncol}{number of columns in the output figure.}

\item{per1000}{logical indicator to multiply results by 1000.}

\item{order}{order of regions when byyear is set to TRUE. Negative values indicate regions are ordered from high to low posterior medians from top to bottom. Positive values indicate from low to high. 0 indicate alphabetic orders.}

\item{...}{additional configurations passed to inla.posterior.sample.}
}
\value{
a data frame of the calculated densities and a ggplot figure.
}
\description{
Function to calculate and plot densities of the projected estimates from INLA for each time and region.
}
\examples{
\dontrun{
years <- levels(DemoData[[1]]$time)

data <- getDirectList(births = DemoData, 
years = years,  
regionVar = "region", timeVar = "time", 
clusterVar = "~clustid+id", 
ageVar = "age", weightsVar = "weights", 
geo.recode = NULL)
# obtain direct estimates
data_multi <- getDirectList(births = DemoData, years = years,
  regionVar = "region",  timeVar = "time", clusterVar = "~clustid+id",
  ageVar = "age", weightsVar = "weights", geo.recode = NULL)
data <- aggregateSurvey(data_multi)

#  national model
years.all <- c(years, "15-19")
fit1 <- smoothDirect(data = data, geo = NULL, Amat = NULL, 
  year_label = years.all, year_range = c(1985, 2019), 
  rw = 2, is.yearly=FALSE, m = 5)
## Plot marginal posterior densities over time
density <- getSmoothedDensity(fit1,  year_label = years.all, year_plot = years.all, 
ncol = 4, byyear = FALSE)
density$g

#  subnational model
fit2 <- smoothDirect(data = data, geo = geo, Amat = mat, 
  year_label = years.all, year_range = c(1985, 2019), 
  rw = 2, is.yearly=TRUE, m = 5, type.st = 1)

# Plot marginal posterior densities over time (regions are ordered alphabetically)
density <- getSmoothedDensity(fit2, Amat = mat, year_label = years.all, 
year_plot = years.all, ncol = 4)
density$g

# Re-order the regions 
density <- getSmoothedDensity(fit2, Amat = mat,  year_plot = years.all,
 ncol = 4, per1000 = TRUE, order = -1)
density$g

# Show each region (instead of each year) in a panel 
## Instead of recalculate the posteriors, we can use previously calculated densities as input 
density <- getSmoothedDensity(results = density, year_plot = years.all, 
ncol = 4, byyear=FALSE, per1000 = TRUE)
density$g

# Show more years
density <- getSmoothedDensity(results = density, year_plot = c(1990:2019), 
ncol = 4, byyear=FALSE, per1000 = TRUE)
density$g



}

}
\seealso{
\code{\link{plot.SUMMERproj}}
}
