% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialAreas.R
\name{getSpatialAreas}
\alias{getSpatialAreas}
\title{Functions for getting spatial polygon areas}
\usage{
getSpatialAreas(mapDat)
}
\arguments{
\item{mapDat}{SpatialPolygonsDataFrame object with map information}
}
\value{
a vector of areas in km^2 of length \code{nrow(mapDat@data)}
}
\description{
Calculates the spatial area, in km^2, of every polygon in the spatialPolygonsDataFrame
}
\examples{
\dontrun{
# download Kenya GADM shapefiles from SUMMERdata github repository
githubURL <- "https://github.com/paigejo/SUMMERdata/blob/main/data/kenyaMaps.rda?raw=true"
tempDirectory = "~/"
mapsFilename = paste0(tempDirectory, "/kenyaMaps.rda")
if(!file.exists(mapsFilename)) {
  download.file(githubURL,mapsFilename)
}

# load it in
out = load(mapsFilename)
out
adm1@data$NAME_1 = as.character(adm1@data$NAME_1)
adm1@data$NAME_1[adm1@data$NAME_1 == "Trans Nzoia"] = "Trans-Nzoia"
adm1@data$NAME_1[adm1@data$NAME_1 == "Elgeyo-Marakwet"] = "Elgeyo Marakwet"
}

}
\author{
John Paige
}
