% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-internals.R
\name{iid.new}
\alias{iid.new}
\title{New random IID models for m-year to period random effects}
\usage{
iid.new(cmd = c("graph", "Q", "mu", "initial", "log.norm.const",
  "log.prior", "quit"), theta = NULL)
}
\arguments{
\item{cmd}{list of model components}

\item{theta}{log precision}
}
\description{
New random IID models for m-year to period random effects
}
