% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenyaMaps.R
\docType{data}
\name{kenyaMaps}
\alias{kenyaMaps}
\alias{adm1Kenya}
\alias{adm2Kenya}
\alias{kenyaPoly}
\alias{provinceMapKenya}
\alias{kenyaMesh}
\title{Maps of Kenya Administrative Areas}
\format{
A number of SpatialPolygonsDataFrame objects and polygons with information about respective administrative areas for mapping. 
Also a spatial triangular mesh over Kenya east/north coordinates from \code{\link{projKenya}} for the SPDE model. The 
dataset names are: adm2Kenya, adm1Kenya, kenyaMesh, kenyaPoly, provinceMapKenya.

An object of class \code{SpatialPolygonsDataFrame} with 273 rows and 79 columns.

An object of class \code{matrix} with 36502 rows and 2 columns.

An object of class \code{SpatialPolygons} of length 8.

An object of class \code{inla.mesh} of length 8.
}
\usage{
data(kenyaMaps)

adm2Kenya

kenyaPoly

provinceMapKenya

kenyaMesh
}
\description{
Maps of Kenya Admin1 and Admin2 along with provinces (Admin1 areas from before the 2009 census)
}
\details{
In the Admin2 level shapefile we combine Admin2 areas in Kenya that have spatial area less than 50 km^2 with 
neighboring ones in the same Admin1 area in order to reduce numerical approximation errors caused 
by representing small areas using the pixel grid, resulting in 273 (possibly combined) Admin2 areas 
in total out of the original 290. The names of the areas that are combined are included in the new 
area name, separated by the "+" symbol.
}
\keyword{datasets}
