% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projKenya.R
\name{kenyaMapUtilities}
\alias{kenyaMapUtilities}
\alias{getAdmin2Kenya}
\alias{getAdmin1Kenya}
\alias{admin2ToAdmin1Kenya}
\title{Utility functions for Kenya administrative areas}
\usage{
getAdmin2Kenya(pts, project = FALSE, delta = 0.05, mean.neighbor = 50)

getAdmin1Kenya(pts, project = FALSE, delta = 0.05, mean.neighbor = 50)

admin2ToAdmin1Kenya(constituencyNames)
}
\arguments{
\item{pts}{2 column matrix of lon/lat (if project == FALSE) or east/north (if project == TRUE) coordinates}

\item{project}{project to longitude/latitude coordinates}

\item{delta}{argument passed to fields.rdist.near in fields package}

\item{mean.neighbor}{argument passed to fields.rdist.near in fields package}

\item{constituencyNames}{character vector of Kenya constituency (Admin1) 
names to convert to county (Admin2) names using adm2Kenya}
}
\description{
Utility functions for Kenya administrative areas
}
\section{Functions}{
\itemize{
\item \code{getAdmin2Kenya}: Calculates Admin2 areas of the input 
spatial coordinates using adm2Kenya variable from kenyaMaps dataset. 
NOTE: this calls \code{data(kenyaMaps)}

\item \code{getAdmin1Kenya}: Computes what administrative regions the given points are in

\item \code{admin2ToAdmin1Kenya}: Gives Kenya Admin1 area containing Admin2 area
}}

\seealso{
\code{\link{projKenya}}
}
\author{
John Paige
}
