% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KingCounty.R
\docType{data}
\name{KingCounty}
\alias{KingCounty}
\title{Map of King County}
\format{An object of class \code{SpatialPolygonsDataFrame} with 48 rows and 9 columns.}
\usage{
KingCounty
}
\description{
Shapefiles are King County in the Washington States.
}
\keyword{datasets}
