% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalawiMap.R
\docType{data}
\name{malawiMap}
\alias{malawiMap}
\title{Malawi map dataset for vignette.}
\format{An object of class \code{SpatialPolygonsDataFrame} with 28 rows and 14 columns.}
\usage{
data(malawiMap)
}
\description{
Shapefiles are from 2015 Kalawi Admin 2 regions provided by DHS. The Likoma island is included in this shapefile. The Admin 2 region names are in the ADM2_EN field.
}
\details{
\itemize{
 \item geo Geographic map files
}
}
\keyword{datasets}
