% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlot.R
\name{mapPlot}
\alias{mapPlot}
\title{Makes map plot}
\usage{
mapPlot(countryname, results, geo, countrysum, inlamod)
}
\arguments{
\item{countryname}{Country name as a string}

\item{results}{output from \code{\link{fit_RW}}}

\item{geo}{\code{geo} output from \code{\link{read_shape}}}

\item{countrysum}{output from \code{\link{countrySummary_mult}}}

\item{inlamod}{output from \code{\link{fitINLA}}}
}
\description{
See package vignette for usage details.
}
