% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_projINLA.R
\name{plot.projINLA}
\alias{plot.projINLA}
\title{Plot projection output.}
\usage{
\method{plot}{projINLA}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{projINLA}}}

\item{...}{optional arguments, see details}
}
\description{
Plot projection output.
}
\details{
Note that arguments after \code{...} must match exactly.
\itemize{
 \item{\code{years_label}}{string of year labels, defaults to \code{c("85-89", "90-94", "95-99", "00-04", "05-09", "10-14", "15-19")}}
 \item{\code{proj_year}}{projection year as numeric, defaults to \code{2015}}
 \item{\code{years_med}}{ median of year intervals, defaults to \code{c(1987, 1992, 1997, 2002, 2007, 2012, 2017)}}
 \item{\code{is.yearly}}{indicator for yearly model, defaults to \code{TRUE}}
 \item{\code{is.subnational}}{indicator for subnational model, defaults to \code{TRUE}}
}
}
\examples{
\dontrun{
TODO
}
}
