% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.SUMMERmodel.svy}
\alias{print.SUMMERmodel.svy}
\title{Print method for the smoothing models from \code{smoothSurvey}.}
\usage{
\method{print}{SUMMERmodel.svy}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{smoothSurvey}}.}

\item{...}{not used}
}
\description{
This function is the print method for class \code{SUMMERmodel.svy}.
}
\examples{
\dontrun{
data(DemoData2)
data(DemoMap2)
fit0 <- smoothSurvey(data=DemoData2,  
Amat=DemoMap2$Amat, responseType="binary", 
responseVar="tobacco.use", strataVar="strata", 
weightVar="weights", regionVar="region", 
clusterVar = "~clustid+id", CI = 0.95)
fit0
}
}
\seealso{
\code{\link{summary.SUMMERmodel.svy}}
}
\author{
Zehang Li
}
