% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projINLA_yearly.R
\name{projINLA}
\alias{projINLA}
\title{Function to obtain projected estimates from INLA for each time and region.}
\usage{
projINLA(inla_mod, is.yearly = TRUE, year_range = c(1985, 2019),
  year_label = c("85-89", "90-94", "95-99", "00-04", "05-09", "10-14",
  "15-19"), Amat = NULL, nsim = 1000)
}
\arguments{
\item{inla_mod}{output from \code{\link{fitINLA}}}

\item{is.yearly}{indicator for whether model is yearly or not}

\item{year_range}{range corresponding to year label}

\item{year_label}{vector of year string vector}

\item{Amat}{adjacency matrix}

\item{nsim}{number of simulations}
}
\value{
Results from RW2 model fit, including projection.
}
\description{
Function to obtain projected estimates from INLA for each time and region.
}
\examples{
\dontrun{
TODO
}


}
