% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projINLA_multi.R
\name{projINLA_multi}
\alias{projINLA_multi}
\title{Function to obtain smoothed estimates from fitted INLA object.}
\usage{
projINLA_multi(fitted, proj.time, ntime, which.area, quantiles,
  return_raw = FALSE)
}
\arguments{
\item{fitted}{Fitted object from \code{\link{fitINLA}}.}

\item{proj.time}{String vector of the time periods to project into the future.}

\item{ntime}{Number of time periods with observation.}

\item{which.area}{String of which area to project.}

\item{quantiles}{Posterior credible interval to return.}

\item{return_raw}{Logical indicator of whether raw posterior draws or summary statistics is returned. Default to FALSE.}
}
\value{
Smoothed model estimates from the provided INLA fit
}
\description{
Function to obtain smoothed estimates from fitted INLA object.
}
\seealso{
\code{\link{countrySummary}}
}
