% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shape.R
\name{read_shape}
\alias{read_shape}
\title{Function to read shape files.}
\usage{
read_shape(filepath, regionnames, data = NULL)
}
\arguments{
\item{filepath}{file path for .shp files}

\item{regionnames}{vector of strings of final region names}

\item{data}{optional country summary data, for checking}
}
\value{
A list including shape files and the adjacency matrix.
}
\description{
Function to read shape files.
}
\examples{
\dontrun{
my_region_names <- c("central","eastern","northern","western")
my_fp <- "myExampleFilepath/sdr_subnational_boundaries.shp"
my_map <- read_shape(filepath = my_fp, regionnames = my_region_names)
}
}
