% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-internal-pc.R
\name{rw.new.pc}
\alias{rw.new.pc}
\title{New random walk 1 and 2 models for m-year to period random effects}
\usage{
rw.new.pc(cmd = c("graph", "Q", "mu", "initial", "log.norm.const",
  "log.prior", "quit"), theta = NULL)
}
\arguments{
\item{cmd}{list of model components}

\item{theta}{log precision}
}
\description{
New random walk 1 and 2 models for m-year to period random effects
}
