% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhyper.R
\name{simhyper}
\alias{simhyper}
\title{Function to simulate hyperpriors from an adjacency matrix.}
\usage{
simhyper(R = 2, nsamp = 1e+05, nsamp.check = 5000, Amat, nperiod = 6)
}
\arguments{
\item{R}{Desired prior odds ratio. Default to 2, i.e., a 95\% prior interval for the residual odds ratios lies in the interval (R, 1/R).}

\item{nsamp}{Sample to simulate for scaling factor}

\item{nsamp.check}{Sample to simulate for checking range}

\item{Amat}{Adjacency matrix of the areas in the data.}

\item{nperiod}{numerical value of how many time periods in the data}
}
\description{
Function to simulate hyperpriors from an adjacency matrix.
}
\examples{
\dontrun{
data(UgandaMap)
mat <- UgandaMap$Amat
priors <- simhyper(R = 2, nsamp = 1e+05, nsamp.check = 5000, Amat = mat)
}
}
\references{
Wakefield, J. Multi-level modelling, the ecologic fallacy, and hybrid study designs. \emph{International Journal of Epidemiology}, 2009, vol. 38 (pg. 330-336).
}
