% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagPlot.R
\name{spagPlot}
\alias{spagPlot}
\title{Makes spaghetti plot}
\usage{
spagPlot(countryname, results, geo, countrysum, inlamod)
}
\arguments{
\item{countryname}{Country name as a string}

\item{results}{output from \code{\link{fit_RW}}}

\item{geo}{geo output from \code{\link{read_shape}}}

\item{countrysum}{output from \code{\link{countrySummary_mult}}}

\item{inlamod}{output from \code{\link{fitINLA}}}
}
\description{
See package vignette for usage details.
}
