% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-internals.R
\name{st.new}
\alias{st.new}
\title{New Type I to IV space time interaction models for m-year to period random effects}
\usage{
st.new(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)
}
\arguments{
\item{cmd}{list of model components}

\item{theta}{log precision}
}
\description{
New Type I to IV space time interaction models for m-year to period random effects
}
