% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMMER-package.R, R/SUMMER.R
\docType{package}
\name{SUMMER-package}
\alias{SUMMER}
\alias{SUMMER-package}
\title{SUMMER: Small-Area-Estimation Unit/Area Models and Methods for Estimation in R}
\description{
Provides methods for spatial and spatio-temporal smoothing of demographic and health indicators using survey data, with particular focus on estimating and projecting under-five mortality rates, described in Mercer et al. (2015) \doi{10.1214/15-AOAS872}, Li et al. (2019) \doi{10.1371/journal.pone.0210645}, Wu et al. (DHS Spatial Analysis Reports No. 21, 2021), and Li et al. (2023) \doi{10.48550/arXiv.2007.05117}.

SUMMER provides methods for spatial and spatio-temporal smoothing of demographic and health indicators using survey data, with particular focus on estimating and projecting under-five mortality rates.
}
\details{
For details on the model implemented in this package, see Mercer et al. (2015) \url{doi:10.1214/15-AOAS872} and Li et al. (2019) \url{doi:10.1371/journal.pone.0210645}.

The development version of the package will be maintained on \url{https://github.com/richardli/SUMMER}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/richardli/SUMMER}
  \item \url{https://richardli.github.io/SUMMER/}
  \item Report bugs at \url{https://github.com/richardli/SUMMER/issues}
}

}
\author{
\strong{Maintainer}: Zehang R Li \email{lizehang@gmail.com}

Authors:
\itemize{
  \item Bryan D Martin \email{bmartin6@uw.edu}
  \item Yuan Hsiao \email{yahsiao@uw.edu}
  \item Jessica Godwin \email{jlg0003@uw.edu}
  \item John Paige \email{paigejo@gmail.com}
  \item Peter Gao \email{petergao@uw.edu}
  \item Jon Wakefield \email{jonno@uw.edu}
  \item Samuel J Clark \email{work@samclark.net}
  \item Geir-Arne Fuglstad \email{geir-arne.fuglstad@ntnu.no}
  \item Andrea Riebler \email{andrea.riebler@ntnu.no}
}

}
\keyword{internal}
