% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uganda.R
\docType{data}
\name{Uganda}
\alias{Uganda}
\title{Fake Uganda dataset for vignette.}
\format{A list of with five components, named by survey year.}
\source{
\url{https://dhsprogram.com/data/model-datasets.cfm}
}
\usage{
Uganda
}
\description{
A small fake dataset with 4 regions and 5 survey years. These numbers do not represent 
any real country's data and are based on a subset of the model dataset
provided by DHS.
}
\keyword{datasets}
