% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UgandaMap.R
\docType{data}
\name{UgandaMap}
\alias{UgandaMap}
\title{Uganda map dataset for vignette.}
\format{An object of class \code{list} of length 2.}
\source{
\url{https://spatialdata.dhsprogram.com/boundaries/#view=table&countryId=UG}
}
\usage{
UgandaMap
}
\description{
Shapefiles are from 1995 Uganda Admin 1 regions provided by DHS.
}
\details{
\itemize{
 \item geo. Geographic map files
 \item Amat. Adjacency matrix for regions
}
}
\keyword{datasets}
