% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeprecatedFunctions.R
\name{fitGeneric}
\alias{fitGeneric}
\title{Deprecated function for smoothed direct estimates for mortality rates}
\usage{
fitGeneric(...)
}
\arguments{
\item{...}{arguments of the new \code{\link{smoothSurvey}} function}
}
\description{
This is an wrapper function for smoothSurvey function for the purpose of backward compatibility.
}
\examples{
\dontrun{
data(DemoData2)
data(DemoMap2)
fit <- smoothSurvey(data=DemoData2,  
Amat=DemoMap2$Amat, responseType="binary", 
responseVar="tobacco.use", strataVar="strata", 
weightVar="weights", regionVar="region", 
clusterVar = "~clustid+id", CI = 0.95)

# Example with region-level covariates
 Xmat <- aggregate(age~region, data = DemoData2, FUN = mean)
 fit <- smoothSurvey(data=DemoData2,  
  Amat=DemoMap2$Amat, responseType="binary", 
  X = Xmat,
  responseVar="tobacco.use", strataVar="strata", 
  weightVar="weights", regionVar="region", 
  clusterVar = "~clustid+id", CI = 0.95)
}
}
\seealso{
\code{\link{smoothSurvey}}
}
