% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeprecatedFunctions.R
\name{fitINLA}
\alias{fitINLA}
\title{Deprecated function for smoothed direct estimates for mortality rates}
\usage{
fitINLA(...)
}
\arguments{
\item{...}{arguments of the new \code{\link{smoothDirect}} function}
}
\description{
This is an wrapper function for smoothDirect function for the purpose of backward compatibility.
}
\examples{
\dontrun{
years <- levels(DemoData[[1]]$time)
# obtain direct estimates
data_multi <- getDirectList(births = DemoData, years = years,
  regionVar = "region",  timeVar = "time", clusterVar = "~clustid+id",
  ageVar = "age", weightsVar = "weights", geo.recode = NULL)
data <- aggregateSurvey(data_multi)

#  national model
years.all <- c(years, "15-19")
fit1 <- smoothDirect(data = data, Amat = NULL, 
  year_label = years.all, year_range = c(1985, 2019), 
  rw = 2, is.yearly=FALSE, m = 5)
out1 <- getSmoothed(fit1)
plot(out1, is.subnational=FALSE)

#  subnational model
fit2 <- smoothDirect(data = data, Amat = DemoMap$Amat, 
  year_label = years.all, year_range = c(1985, 2019), 
  rw = 2, is.yearly=TRUE, m = 5, type.st = 4)
out2 <- getSmoothed(fit2, Amat = mat)
plot(out2, is.yearly=TRUE, is.subnational=TRUE)
}
}
\seealso{
\code{\link{smoothDirect}}
}
