% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitINLA.R
\name{fitINLA}
\alias{fitINLA}
\title{Fit space-time smoothing models to mortality rates}
\usage{
fitINLA(data, Amat, geo, formula = NULL, rw = 2, is.yearly = TRUE,
  year_names, year_range = c(1980, 2014), m = 5, na.rm = TRUE,
  priors = NULL, type.st = 1, hyper = c("pc", "gamma")[1],
  pc.u = 1, pc.alpha = 0.01, pc.u.phi = 0.5, pc.alpha.phi = 2/3,
  a.iid = NULL, b.iid = NULL, a.rw = NULL, b.rw = NULL,
  a.icar = NULL, b.icar = NULL, options = list(dic = T, mlik = T, cpo
  = T, openmp.strategy = "default"), verbose = FALSE)
}
\arguments{
\item{data}{Combined dataset}

\item{Amat}{Adjacency matrix for the regions}

\item{geo}{Geo file}

\item{formula}{INLA formula. Defaults to RW2, ICAR, IID time, IID, region, IID survey effect, IID time-region interaction, IID survey-region interaction, and IID survey-time-region interaction.}

\item{rw}{Take values 1 or 2, indicating the order of random walk.}

\item{is.yearly}{Logical indicator for fitting yearly or period model.}

\item{year_names}{string vector of year names}

\item{year_range}{Entire range of the years (inclusive) defined in year_names.}

\item{m}{Number of years in each period.}

\item{na.rm}{Logical indicator of whether to remove rows with NA values in the data. Default set to TRUE.}

\item{priors}{priors from \code{\link{simhyper}}}

\item{type.st}{type for space-time interaction}

\item{hyper}{option to choose between PC prior or Gamma prior for the latent factors.}

\item{a.iid}{hyperparameter for i.i.d random effects.}

\item{b.iid}{hyperparameter for i.i.d random effects.}

\item{a.rw}{hyperparameter for RW 1 or 2 random effects.}

\item{b.rw}{hyperparameter for RW 1 or 2random effects.}

\item{a.icar}{hyperparameter for ICAR random effects.}

\item{b.icar}{hyperparameter for ICAR random effects.}

\item{options}{list of options to be passed to control.compute() in the inla() function.}

\item{verbose}{logical indicator to print out detailed inla() intermediate steps.}
}
\value{
INLA model fit using the provided formula, country summary data, and geographic data
}
\description{
Fit space-time smoothing models to mortality rates
}
\examples{
\dontrun{

data(DemoData)
data(DemoMap)
years <- levels(DemoData[[1]]$time)

# obtain direct estimates
data <- getDirectList(births = DemoData, 
years = years, 
regionVar = "region", timeVar = "time", 
clusterVar = "~clustid+id", 
ageVar = "age", weightsVar = "weights", 
geo.recode = NULL)

# obtain maps
geo <- DemoMap$geo
mat <- DemoMap$Amat

# Simulate hyperpriors
priors <- simhyper(R = 2, nsamp = 1e+05, nsamp.check = 5000, Amat = mat, only.iid = TRUE)

# combine data from multiple surveys
data <- aggregateSurvey(data)

# Model fitting with INLA
years.all <- c(years, "15-19")
fit <- fitINLA(data = data, geo = geo, Amat = mat, 
year_names = years.all, year_range = c(1985, 2019),
 priors = priors, rw = 2,
 is.yearly=TRUE, m = 5, type.st = 4)
# Projection
out <- getSmoothed(fit, Amat = mat, is.yearly = TRUE)
plot(out, is.yearly=TRUE, is.subnational=TRUE) + ggplot2::ggtitle("Subnational yearly model")

}

}
\seealso{
\code{\link{getDirect}}
}
