% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitINLA.R
\name{fitINLA}
\alias{fitINLA}
\title{Fit space-time smoothed direct models to mortality rates}
\usage{
fitINLA(data, Amat, X = NULL, formula = NULL, rw = 2, ar = 0,
  is.yearly = TRUE, year_label, year_range = c(1980, 2014), m = 5,
  na.rm = TRUE, priors = NULL, type.st = 1, survey.effect = FALSE,
  hyper = c("pc", "gamma")[1], pc.u = 1, pc.alpha = 0.01,
  pc.u.phi = 0.5, pc.alpha.phi = 2/3, pc.u.cor = 0.7,
  pc.alpha.cor = 0.9, a.iid = NULL, b.iid = NULL, a.rw = NULL,
  b.rw = NULL, a.icar = NULL, b.icar = NULL, options = list(dic =
  T, mlik = T, cpo = T, openmp.strategy = "default"),
  control.inla = list(strategy = "adaptive", int.strategy = "auto"),
  verbose = FALSE, pc.st.u = NA, pc.st.alpha = NA, geo = NULL)
}
\arguments{
\item{data}{Combined dataset}

\item{Amat}{Adjacency matrix for the regions}

\item{X}{Covariate matrix. It must contain either a column with name "region", or a column with name "years", or both. The covariates must not have missing values for all regions (if varying in space) and all time periods (if varying in time). The rest of the columns are treated as covariates in the mean model.}

\item{formula}{INLA formula. See vignette for example of using customized formula.}

\item{rw}{Take values 0, 1 or 2, indicating the order of random walk. If rw = 0, the autoregressive process is used instead of the random walk in the main trend. See the description of the argument ar for details.}

\item{ar}{Order of the autoregressive component. If ar is specified to be positive integer, the random walk components will be replaced by AR(p) terms in the interaction part. The main temporal trend remains to be random walk of order rw unless rw = 0.}

\item{is.yearly}{Logical indicator for fitting yearly or period model.}

\item{year_label}{string vector of year names}

\item{year_range}{Entire range of the years (inclusive) defined in year_label.}

\item{m}{Number of years in each period.}

\item{na.rm}{Logical indicator of whether to remove rows with NA values in the data. Default set to TRUE.}

\item{priors}{priors from \code{\link{simhyper}}}

\item{type.st}{type for space-time interaction}

\item{survey.effect}{logical indicator whether to include a survey iid random effect. If this is set to TRUE, there needs to be a column named 'survey' in the input data frame. In prediction, this random effect term will be set to 0.}

\item{hyper}{which hyperpriors to use. Default to be using the PC prior ("pc").}

\item{pc.u}{hyperparameter U for the PC prior on precisions.}

\item{pc.alpha}{hyperparameter alpha for the PC prior on precisions.}

\item{pc.u.phi}{hyperparameter U for the PC prior on the mixture probability phi in BYM2 model.}

\item{pc.alpha.phi}{hyperparameter alpha for the PC prior on the mixture probability phi in BYM2 model.}

\item{pc.u.cor}{hyperparameter U for the PC prior on the autocorrelation parameter in the AR prior, i.e. Prob(cor > pc.u.cor) = pc.alpha.cor.}

\item{pc.alpha.cor}{hyperparameter alpha for the PC prior on the autocorrelation parameter in the AR prior.}

\item{a.iid}{hyperparameter for i.i.d random effects.}

\item{b.iid}{hyperparameter for i.i.d random effects.}

\item{a.rw}{hyperparameter for RW 1 or 2 random effects.}

\item{b.rw}{hyperparameter for RW 1 or 2random effects.}

\item{a.icar}{hyperparameter for ICAR random effects.}

\item{b.icar}{hyperparameter for ICAR random effects.}

\item{options}{list of options to be passed to control.compute() in the inla() function.}

\item{control.inla}{list of options to be passed to control.inla() in the inla() function. Default to the "adaptive" integration strategy.}

\item{verbose}{logical indicator to print out detailed inla() intermediate steps.}

\item{pc.st.u}{hyperparameter U for the PC prior on precisions for the interaction term.}

\item{pc.st.alpha}{hyperparameter alpha for the PC prior on precisions for the interaction term.}

\item{geo}{Spatial polygon file, legacy parameter from previous versions of the package.}
}
\value{
INLA model fit using the provided formula, country summary data, and geographic data
}
\description{
Fit space-time smoothed direct models to mortality rates
}
\examples{
\dontrun{
years <- levels(DemoData[[1]]$time)

# obtain direct estimates
data <- getDirectList(births = DemoData, 
years = years,  
regionVar = "region", timeVar = "time", 
clusterVar = "~clustid+id", 
ageVar = "age", weightsVar = "weights", 
geo.recode = NULL)
# obtain direct estimates
data_multi <- getDirectList(births = DemoData, years = years,
  regionVar = "region",  timeVar = "time", clusterVar = "~clustid+id",
  ageVar = "age", weightsVar = "weights", geo.recode = NULL)
data <- aggregateSurvey(data_multi)

#  national model
years.all <- c(years, "15-19")
fit1 <- fitINLA(data = data, Amat = NULL, 
  year_label = years.all, year_range = c(1985, 2019), 
  rw = 2, is.yearly=FALSE, m = 5)
out1 <- getSmoothed(fit1)
plot(out1, is.subnational=FALSE)

#  subnational model
fit2 <- fitINLA(data = data, Amat = mat, 
  year_label = years.all, year_range = c(1985, 2019), 
  rw = 2, is.yearly=TRUE, m = 5, type.st = 4)
out2 <- getSmoothed(fit2, Amat = mat)
plot(out2, is.yearly=TRUE, is.subnational=TRUE)


}
}
\seealso{
\code{\link{getDirect}}
}
