% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitINLA2.R
\name{fitINLA2}
\alias{fitINLA2}
\title{Fit cluster-level space-time smoothing models to mortality rates}
\usage{
fitINLA2(data, family = c("betabinomial", "binomial")[1],
  age.groups = c("0", "1-11", "12-23", "24-35", "36-47", "48-59"),
  age.n = c(1, 11, 12, 12, 12, 12), Amat, geo, bias.adjust = NULL,
  formula = NULL, rw = 2, is.yearly = TRUE, year_names,
  year_range = c(1980, 2014), m = 5, na.rm = TRUE, priors = NULL,
  type.st = 1, hyper = c("pc", "gamma")[1], pc.u = 1,
  pc.alpha = 0.01, pc.u.phi = 0.5, pc.alpha.phi = 2/3,
  a.iid = NULL, b.iid = NULL, a.rw = NULL, b.rw = NULL,
  a.icar = NULL, b.icar = NULL, options = list(config = TRUE),
  verbose = FALSE)
}
\arguments{
\item{data}{count data of person-months with the following columns
\itemize{
 \item cluster: cluster ID
 \item years: time period
 \item region: region of the cluster
\item strata: stratum of the cluster
\item age: age group corresponding to the row
\item total: total number of person-month in this age group, stratum, cluster, and period
\item Y: total number of deaths in this age group, stratum, cluster, and period
}}

\item{family}{family of the model. This can be either binomial (with logistic normal prior) or betabiniomial.}

\item{age.groups}{a character vector of age groups in increasing order.}

\item{age.n}{number of months in each age groups in the same order.}

\item{Amat}{Adjacency matrix for the regions}

\item{geo}{Geo file}

\item{bias.adjust}{the ratio of unadjusted mortality rates or age-group-specific hazards to the true rates or hazards. It needs to be a data frame that can be merged to thee outcome, i.e., with the same column names for time periods (for national adjustment), or time periods and region (for subnational adjustment). The column specifying the ratio should be named "ratio".}

\item{formula}{INLA formula.  See vignette for example of using customized formula.}

\item{rw}{Take values 1 or 2, indicating the order of random walk.}

\item{is.yearly}{Logical indicator for fitting yearly or period model.}

\item{year_names}{string vector of year names}

\item{year_range}{Entire range of the years (inclusive) defined in year_names.}

\item{m}{Number of years in each period.}

\item{na.rm}{Logical indicator of whether to remove rows with NA values in the data. Default set to TRUE.}

\item{priors}{priors from \code{\link{simhyper}}}

\item{type.st}{type for space-time interaction}

\item{hyper}{which hyperpriors to use. Default to be using the PC prior ("pc").}

\item{pc.u}{hyperparameter U for the PC prior on precisions.}

\item{pc.alpha}{hyperparameter alpha for the PC prior on precisions.}

\item{pc.u.phi}{hyperparameter U for the PC prior on the mixture probability phi in BYM2 model.}

\item{pc.alpha.phi}{hyperparameter alpha for the PC prior on the mixture probability phi in BYM2 model.}

\item{a.iid}{hyperparameter for i.i.d random effects.}

\item{b.iid}{hyperparameter for i.i.d random effects.}

\item{a.rw}{hyperparameter for RW 1 or 2 random effects.}

\item{b.rw}{hyperparameter for RW 1 or 2random effects.}

\item{a.icar}{hyperparameter for ICAR random effects.}

\item{b.icar}{hyperparameter for ICAR random effects.}

\item{options}{list of options to be passed to control.compute() in the inla() function.}

\item{verbose}{logical indicator to print out detailed inla() intermediate steps.}
}
\value{
INLA model fit using the provided formula, country summary data, and geographic data
}
\description{
Fit cluster-level space-time smoothing models to mortality rates
}
\examples{
\dontrun{
 

}

}
\seealso{
\code{\link{getDirect}}
}
