% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeprecatedFunctions.R
\name{fitINLA2}
\alias{fitINLA2}
\title{Deprecated function for cluster-level space-time smoothing models for mortality rates}
\usage{
fitINLA2(...)
}
\arguments{
\item{...}{arguments of the new \code{\link{smoothCluster}} function}
}
\description{
This is an wrapper function for smoothCluster function for the purpose of backward compatibility.
}
\examples{
\dontrun{
library(dplyr)
data(DemoData)
# Create dataset of counts
counts.all <- NULL
for(i in 1:length(DemoData)){
  counts <- getCounts(DemoData[[i]][, c("clustid", "time", "age", "died",
                                        "region", "strata")],
            variables = 'died', by = c("age", "clustid", "region", 
                                         "time", "strata"))
  counts <- counts \%>\% mutate(cluster = clustid, years = time, Y=died)
  counts$strata <- gsub(".*\\\\.","",counts$strata)
  counts$survey <- names(DemoData)[i] 
  counts.all <- rbind(counts.all, counts)
}

# fit cluster-level model on the periods
periods <- levels(DemoData[[1]]$time)
fit <- smoothCluster(data = counts.all, 
      Amat = DemoMap$Amat, 
      time.model = "rw2", 
      st.time.model = "rw1",
      strata.time.effect =  TRUE, 
      survey.effect = TRUE,
      family = "betabinomial",
      year_label = c(periods, "15-19"))
est <- getSmoothed(fit, Amat = DemoMap$Amat, 
      year_label =c(periods, "15-19"), nsim = 1000)
plot(est$stratified, plot.CI=TRUE) + ggplot2::facet_wrap(~strata) 
}
}
\seealso{
\code{\link{smoothCluster}}
}
