% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBirths.R
\name{getBirths}
\alias{getBirths}
\title{Function to get Births file from DHS .dta files.}
\usage{
getBirths(filepath, surveyyear)
}
\arguments{
\item{filepath}{file path of raw .dta file from DHS}

\item{surveyyear}{year of survey}
}
\value{
A list of birth-month data
}
\description{
Function to get Births file from DHS .dta files.
}
\examples{
\dontrun{
my_fp <- "/myExampleFilepath/surveyData.DTA"
Uganda <- getBirths(filepath = my_fp, surveyyear = 2015) 
}

}
