% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiag.R
\name{getDiag}
\alias{getDiag}
\title{Make diagnostic plots}
\usage{
getDiag(inla_mod, field = c("space", "time", "spacetime")[1],
  Amat = NULL, CI = 0.95, draws = NULL, ...)
}
\arguments{
\item{inla_mod}{output from \code{\link{smoothDirect}}}

\item{field}{which random effects to plot. It can be one of the following: space, time, and spacetime.}

\item{Amat}{adjacency matrix}

\item{CI}{Desired level of credible intervals}

\item{draws}{Posterior samples drawn from the fitted model. This argument allows the previously sampled draws (by setting save.draws to be TRUE) be used in new aggregation tasks.}

\item{...}{unused arguments}
}
\value{
List of diagnostic plots
}
\description{
Make diagnostic plots
}
\examples{
\dontrun{
  data(DemoMap)
  years <- levels(DemoData[[1]]$time)
  
  # obtain direct estimates
  data <- getDirectList(births = DemoData, 
  years = years,  
  regionVar = "region", timeVar = "time", 
  clusterVar = "~clustid+id", 
  ageVar = "age", weightsVar = "weights", 
  geo.recode = NULL)
  # obtain direct estimates
  data_multi <- getDirectList(births = DemoData, years = years,
    regionVar = "region",  timeVar = "time", clusterVar = "~clustid+id",
    ageVar = "age", weightsVar = "weights", geo.recode = NULL)
  data <- aggregateSurvey(data_multi)
  
  #  national model
  years.all <- c(years, "15-19")
  fit1 <- smoothDirect(data = data, geo = DemoMap$geo, Amat = DemoMap$Amat, 
    year_label = years.all, year_range = c(1985, 2019), 
    rw = 2, is.yearly=FALSE, m = 5)
random.time <- getDiag(fit1, field = "time")
  random.space <- getDiag(fit1, field = "space", Amat = DemoMap$Amat)
  random.spacetime <- getDiag(fit1, field = "spacetime",
   Amat = DemoMap$Amat)
}

}
