% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalawiMap.R
\docType{data}
\name{MalawiMap}
\alias{MalawiMap}
\title{Malawi Admin-2 map}
\format{An object of class \code{SpatialPolygonsDataFrame} with 28 rows and 14 columns.}
\usage{
data(MalawiMap)
}
\description{
SpatialPolygonsDataFrame objects that reflect the Admin 2 regions in Malawi, including the Likoma island. The Admin 2 region names are in the ADM2_EN field.
}
\keyword{datasets}
