% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmoothedDensity.R
\name{ridgePlot}
\alias{ridgePlot}
\alias{getSmoothedDensity}
\title{Calculate and plot posterior densities of the projected estimates}
\usage{
ridgePlot(
  x = NULL,
  nsim = 1000,
  draws = NULL,
  Amat = NULL,
  year_plot = NULL,
  strata_plot = NULL,
  by.year = TRUE,
  ncol = 4,
  scale = 2,
  per1000 = FALSE,
  order = 0,
  direction = 1,
  results = NULL,
  ...
)

getSmoothedDensity(
  x = NULL,
  nsim = 1000,
  draws = NULL,
  Amat = NULL,
  year_plot = NULL,
  strata_plot = NULL,
  by.year = TRUE,
  ncol = 4,
  scale = 2,
  per1000 = FALSE,
  order = 0,
  direction = 1,
  results = NULL,
  ...
)
}
\arguments{
\item{x}{output from \code{\link{smoothDirect}} for the smoothed direct estimates, or \code{\link{smoothCluster}} for the cluster-level estimates.}

\item{nsim}{number of posterior draws to take. Only used for cluster-level models when \code{draws} is NULL. Otherwise the posterior draws in \code{draws} will be used instead without resampling.}

\item{draws}{Output of \code{\link{getSmoothed}} with \code{save.draws} set to TRUE. This argument allows the previously sampled draws (by setting \code{save.draws} to be TRUE) be used in new aggregation tasks. This argument is only used for cluster-level models.}

\item{Amat}{adjacency matrix}

\item{year_plot}{A vector indicate which years to plot}

\item{strata_plot}{Name of the strata to plot. If not specified, the overall is plotted.}

\item{by.year}{logical indicator for whether the output uses years as facets.}

\item{ncol}{number of columns in the output figure.}

\item{scale}{numerical value controlling the height of the density plots.}

\item{per1000}{logical indicator to multiply results by 1000.}

\item{order}{order of regions when by.year is set to TRUE. Negative values indicate regions are ordered from high to low posterior medians from top to bottom. Positive values indicate from low to high. 0 indicate alphabetic orders.}

\item{direction}{Direction of the color scheme. It can be either 1 (smaller values are darker) or -1 (higher values are darker). Default is set to 1.}

\item{results}{output from \code{\link{ridgePlot}}. This argument can be specified to avoid calculating densities again when only the visualization changes.}

\item{...}{additional configurations passed to inla.posterior.sample.}
}
\value{
a data frame of the calculated densities and a ggplot figure.
}
\description{
The function \code{ridgePlot} replaces the previous function name \code{getSmoothedDensity} (before version 1.0.0).
}
\examples{
\dontrun{
years <- levels(DemoData[[1]]$time)

data <- getDirectList(births = DemoData, 
years = years,  
regionVar = "region", timeVar = "time", 
clusterVar = "~clustid+id", 
ageVar = "age", weightsVar = "weights", 
geo.recode = NULL)
# obtain direct estimates
data_multi <- getDirectList(births = DemoData, years = years,
  regionVar = "region",  timeVar = "time", clusterVar = "~clustid+id",
  ageVar = "age", weightsVar = "weights", geo.recode = NULL)
data <- aggregateSurvey(data_multi)

#  national model
years.all <- c(years, "15-19")
fit1 <- smoothDirect(data = data, geo = NULL, Amat = NULL, 
  year_label = years.all, year_range = c(1985, 2019), 
  rw = 2, is.yearly=FALSE, m = 5)
## Plot marginal posterior densities over time
density <- ridgePlot(fit1, year_plot = years.all, 
ncol = 4, by.year = FALSE)
density$g

#  subnational model
fit2 <- smoothDirect(data = data, geo = geo, Amat = mat, 
  year_label = years.all, year_range = c(1985, 2019), 
  rw = 2, is.yearly=TRUE, m = 5, type.st = 1)

# Plot marginal posterior densities over time (regions are ordered alphabetically)
density <- ridgePlot(fit2, Amat = mat, year_plot = years.all, ncol = 4)
density$g

# Re-order the regions 
density <- ridgePlot(fit2, Amat = mat,  year_plot = years.all,
 ncol = 4, per1000 = TRUE, order = -1)
density$g

# Show each region (instead of each year) in a panel 
## Instead of recalculate the posteriors, we can use previously calculated densities as input 
density <- ridgePlot(results = density, year_plot = years.all, 
ncol = 4, by.year=FALSE, per1000 = TRUE)
density$g

# Show more years
density <- ridgePlot(results = density, year_plot = c(1990:2019), 
ncol = 4, by.year=FALSE, per1000 = TRUE)
density$g



}

}
\seealso{
\code{\link{plot.SUMMERproj}}
}
\author{
Zehang Richard Li
}
