% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPop.R
\name{simPopInternal}
\alias{simPopInternal}
\alias{getExpectedNperEA}
\alias{getSortIndices}
\alias{rStratifiedMultnomial}
\alias{rStratifiedMultnomialBySubarea}
\alias{rMyMultinomial}
\alias{rMyMultinomialSubarea}
\alias{rmultinom1}
\alias{sampleNMultilevelMultinomial}
\alias{sampleNMultilevelMultinomialFixed}
\title{Internal functions for population simulation}
\usage{
getExpectedNperEA(easpa, popMat)

getSortIndices(
  i,
  urban = TRUE,
  popMat,
  stratifyByUrban = TRUE,
  validationPixelI = NULL
)

rStratifiedMultnomial(n, popMat, easpa, stratifyByUrban = TRUE)

rStratifiedMultnomialBySubarea(
  n,
  popMat,
  easpa,
  stratifyByUrban = TRUE,
  poppsub = NULL,
  min1PerSubarea = TRUE,
  minSample = 1
)

rMyMultinomial(
  n,
  i,
  stratifyByUrban = TRUE,
  urban = TRUE,
  popMat = NULL,
  easpa = NULL,
  min1PerSubarea = FALSE,
  method = c("mult1", "mult", "indepMH"),
  minSample = 1
)

rMyMultinomialSubarea(
  n,
  i,
  easpsub,
  stratifyByUrban = TRUE,
  urban = TRUE,
  popMat = NULL
)

rmultinom1(
  n = 1,
  size,
  prob,
  maxSize = 5000 * 5000,
  method = c("mult1", "mult", "indepMH"),
  verbose = FALSE,
  minSample = 100,
  maxExpectedSizeBeforeSwitch = 1000 * 1e+07,
  init = NULL,
  burnIn = floor(n/4),
  filterEvery = 10,
  zeroProbZeroSamples = TRUE,
  allowSizeLessThanK = FALSE
)

sampleNMultilevelMultinomial(
  nDraws = ncol(pixelIndexMat),
  pixelIndexMat = NULL,
  urbanMat = NULL,
  areaMat = NULL,
  easpaList,
  popMat,
  stratifyByUrban = TRUE,
  verbose = TRUE,
  returnEAinfo = FALSE
)

sampleNMultilevelMultinomialFixed(
  clustersPerPixel,
  nDraws = ncol(pixelIndices),
  pixelIndices = NULL,
  urbanVals = NULL,
  areaVals = NULL,
  easpa,
  popMat,
  stratifyByUrban = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{easpa}{census frame. See \code{\link{simPopPixel}} for details}

\item{popMat}{data.frame of pixellated grid of population densities. See \code{\link{simPopPixel}} for details}

\item{i}{index}

\item{urban}{if TRUE, calculate only for urban part of the area. If FALSE, for only rural part}

\item{stratifyByUrban}{whether or not to stratify calculations by urban/rural classification}

\item{validationPixelI}{CURRENTLY FOR TESTING PURPOSES ONLY a set of indices of pixels for which we want to simulate populations (used for pixel level validation)}

\item{n}{number of samples}

\item{poppsub}{population per subarea. See \code{\link{simPopPixel}} for details}

\item{min1PerSubarea}{whether or not to ensure there is at least 1 EA per subarea. See \code{\link{simPopPixel}} for details}

\item{minSample}{the minimum number of samples per `chunk` of samples for truncated multinomial sampling. Defaults to 1}

\item{method}{if min1PerSubarea is TRUE, the sampling method for the truncated multinomial to use with rmulitnom1. rmultinom1 automatically 
        switches between them depending on the number of expected samples. The methods are:
\describe{
  \item{mult1}{rejection sampling from multinomial plus 1 in each category}
  \item{mult}{rejection sampling from multinomial if any category has zero count}
  \item{indepMH}{independent Metropolis-Hastings using multinomial plus 1 distribution as proposal}
}}

\item{easpsub}{this could either be total EAs per subarea, or subarea crossed with urban or 
rural if stratifyByUrban is TRUE}

\item{size}{multinomial size parameter. See \code{\link[stats]{rmultinom}}}

\item{prob}{multinomial probability vector parameter. See \code{\link[stats]{rmultinom}}}

\item{maxSize}{the maximum number of elements in a matrix drawn from the proposal distribution per sample chunk.}

\item{verbose}{whether to print progress as the function proceeds}

\item{maxExpectedSizeBeforeSwitch}{max expected number of samples / k, the number of categories, before switching method}

\item{init}{initial sample if method is `indepMH`}

\item{burnIn}{number of initial samples before samples are collected if method is `indepMH`}

\item{filterEvery}{store only every filterEvery samples if method is i`indepMH`}

\item{zeroProbZeroSamples}{if TRUE, set samples for parts of prob vector that are zero to zero. Otherwise they are set to one.}

\item{allowSizeLessThanK}{if TRUE, then if size < the number of categories (k), returns matrix where each 
column is vector of size ones and k - size zeros. If FALSE, throws an error if size < k}

\item{nDraws}{number of draws}

\item{pixelIndexMat}{matrix of pixel indices associated with each EA and draw}

\item{urbanMat}{matrix of urbanicities associated with each EA and draw}

\item{areaMat}{matrix of areas associated with each EA and draw}

\item{easpaList}{a list of length n with each element being of the format of easpa 
giving the number of households and EAs 
per stratum. It is assumed that the number of EAs per stratum is 
the same in each list element. If easpaList is a data frame, 
number of households per stratum is assumed constant}

\item{returnEAinfo}{whether a data frame at the EA level is desired}

\item{clustersPerPixel}{CURRENTLY FOR TESTING PURPOSES ONLY a vector of length nPixels specifying the number of clusters per pixel if they are fixed}

\item{pixelIndices}{a nEA x n matrix of pixel indices associated with each EA per simulation/draw}

\item{urbanVals}{a nEA x n matrix of urbanicities associated with each EA per simulation/draw}

\item{areaVals}{a nEA x n matrix of area names associated with each EA per simulation/draw}
}
\description{
Functions for calculating valuable quantities and for drawing from important 
distributions for population simulation.
}
\section{Functions}{
\itemize{
\item \code{getExpectedNperEA}: Calculates expected denominator per enumeration area

\item \code{getSortIndices}: For recombining separate multinomials into the draws over all pixels

\item \code{rStratifiedMultnomial}: Gives nPixels x n matrix of draws from the stratified multinomial with values 
corresponding to the value of |C^g| for each pixel, g (the number of EAs/pixel)

\item \code{rStratifiedMultnomialBySubarea}: Gives nPixels x n matrix of draws from the stratified multinomial with values

\item \code{rMyMultinomial}: 

\item \code{rMyMultinomialSubarea}: 

\item \code{rmultinom1}: Random (truncated) multinomial draws conditional on the number of each type being at least one

\item \code{sampleNMultilevelMultinomial}: Take multilevel multinomial draws first from joint distribution of 
number of households per EA given the total per stratum, and then from the joint 
distribution of the total target population per household given 
the total per stratum

\item \code{sampleNMultilevelMultinomialFixed}: Same as sampleNMultilevelMultinomial, except the number of EAs per pixel is fixed
}}

