% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMMER.R
\name{SUMMER-package}
\alias{SUMMER-package}
\alias{SUMMER}
\title{SUMMER package documentation.}
\description{
SUMMER provides implementation for Spatio-temporal Under-five Mortality Methods for Estimation.
}
\details{
The package will be maintained on \url{https://github.com/martinbryan/SUMMER}.
}
