% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMMER.R
\name{SUMMER-package}
\alias{SUMMER-package}
\alias{SUMMER}
\title{SUMMER package documentation.}
\description{
SUMMER provides methods for spatial and spatio-temporal smoothing of demographic and health indicators using survey data, with particular focus on estimating and projecting under-five mortality rates.
}
\details{
For details on the model implemented in this package, see Mercer et al. (2015) <doi:10.1214/15-AOAS872> and Li et al. (2019) <doi:10.1371/journal.pone.0210645>.

The development version of the package will be maintained on \url{https://github.com/richardli/SUMMER}.
}
