% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMMER.R
\name{SUMMER-package}
\alias{SUMMER-package}
\alias{SUMMER}
\title{SUMMER package documentation.}
\description{
SUMMER provides methods for estimating, projecting, and plotting spatio-temporal under-five mortality rates.
}
\details{
For details on the model implemented in this package, see Mercer et al. (2015) <doi:10.1214/15-AOAS872>.

The development version of the package will be maintained on \url{https://github.com/martinbryan/SUMMER}.
}
